q1q2

+++ function-1-1.cpp
@@ -8,7 +8,7 @@ using namespace std;
 
 double arrayMin(double* array, int size)
 {
-    int min = array[0];
+    double min = array[0];
     for (int i = 1; i < size; i++)
     {
         if (min > array[i])
